- Check if chanfix scores are too low before the fix is added. - Re-order add_chanfix() to avoid a couple of unnecessary malloc() and   free() calls. - Add a couple of add NOTE calls to give more channel info.
git-svn-id: http://svn.ratbox.org/svnroot/rserv/trunk@27151 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/s_chanfix.h
@@ -48,8 +48,8 @@ struct chanfix_score_item
 {
 	struct chmember *msptr;
 	unsigned long userhost_id;
-	int score;
-	uint8_t opped;
+	unsigned short score;
+	int8_t opped;
 };
 
 /* The minimum number of ops required for a successful chanfix. */
